+++ src/scripts/objects/player.js
@@ -1,0 +1,70 @@
+/* global Phaser */
+
+import 'phaser';
+import Entity from './entity';
+import PlayerLaser from './playerLaser';
+
+export default class Player extends Entity {
+  constructor(scene, x, y, key) {
+    super(scene, x, y, key, 'Player');
+    this.setData('speed', 200);
+    this.play('dodo');
+
+    this.setData('isShooting', false);
+    this.setData('timerShootDelay', 10);
+    this.setData('timerShootTick', this.getData('timerShootDelay') - 1);
+  }
+
+  moveUp() {
+    this.body.velocity.y = -this.getData('speed');
+  }
+
+  moveDown() {
+    this.body.velocity.y = this.getData('speed');
+  }
+
+  moveLeft() {
+    this.body.velocity.x = -this.getData('speed');
+  }
+
+  moveRight() {
+    this.body.velocity.x = this.getData('speed');
+  }
+
+  onDestroy() {
+    this.scene.time.addEvent({ // go to game over scene
+      delay: 1000,
+      callback() {
+        const { globals } = this.scene.sys.game;
+        if (this.scene.score > globals.model.score.score) {
+          globals.model.score = { score: this.scene.score, user: globals.model.score.user };
+          globals.model.save();
+        }
+        globals.currentScore = this.scene.score;
+        this.scene.score = 1;
+        this.scene.scene.start('GameOver');
+      },
+      callbackScope: this,
+      loop: false,
+    });
+  }
+
+  update() {
+    this.body.setVelocity(0, 0);
+
+    this.x = Phaser.Math.Clamp(this.x, 0, this.scene.game.config.width);
+    this.y = Phaser.Math.Clamp(this.y, 0, this.scene.game.config.height);
+
+    if (this.getData('isShooting')) {
+      if (this.getData('timerShootTick') < this.getData('timerShootDelay')) {
+        this.setData('timerShootTick', this.getData('timerShootTick') + 1); // every game update, increase timerShootTick by one until we reach the value of timerShootDelay
+      } else { // when the "manual timer" is triggered:
+        const laser = new PlayerLaser(this.scene, this.x, this.y);
+        this.scene.playerLasers.add(laser);
+
+        this.scene.sfx.laser.play(); // play the laser sound effect
+        this.setData('timerShootTick', 0);
+      }
+    }
+  }
+}
